% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_gr.R
\docType{data}
\name{snp_gr}
\alias{snp_gr}
\title{Single Nucleotide Polymorphisms GRanges Object}
\format{
A \code{GRanges} object with 30,000 SNPs.
}
\usage{
data(snp_gr)
}
\description{
This dataset contains a \code{GRanges} object storing single nucleotide 
polymorphisms (SNPs).
}
\keyword{datasets}
