% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_gr.R
\docType{data}
\name{snv_gr}
\alias{snv_gr}
\title{Single Nucleotide Variant GRanges Object}
\format{
A \code{GRanges} object with 50,000 SNVs.
}
\usage{
data(snv_gr)
}
\description{
This dataset contains a \code{GRanges} object storing single nucleotide 
variants (SNVs).
}
\keyword{datasets}
